chore: initial commit
This is the initial commit for the repository. It sets up some basic
ideas about the bootstrap theme and gets things to a point where maybe
we can envision where to take things from here.

Bootstrap is pulled in via SCSS source so that we can compile it and set
variables and what-not. There's quite a bit more that is hardcoded into
the _layouts/ and _includes/ than there should be but overall, this is a
start.