feat: stub new home page design
This **does not** totally update the site design. The Cover still needs
a lot of work, backgrounds need to be added, and spacing/sizing needs
improvement. But this is a start! This has the content, most of the
colors, and the general layout down. It also (hopefully) makes the SCSS
a bit more customizable.

Events are updated to be an `include` that accept an event as a
parameter. This allows us to reuse them between the events page and the
home page of the site. We only show the next 3 events on the home page.