feat: automatically remove past events
Okay... this isn't quite "automatic". The past events will still exist
until the next time the site is built. This is why the site is also now
built automatically at least once per day over night. This is actually a
semi-nice compromise for now; visitors may want to see an event that
happened the same day.

Along with this, long event descriptions are truncated and when that
happens, a link to the full event details page is provided. If the
details are not truncated, that isn't done because it provides little
value over the card alone (but the page does still technically).

This page is currently the only way to reference an event after it has
occurred.