doc(README): add notes on timestamps for events
The January meetup was shown as being at 4:30PM ET due to the UTC offset
in the front matter being `-04:00` while the current east coast time is
`-05:00`. This corrects that for this event (even though it's over) and
adds a note to the README and the `event-frontmatter.txt` template file
about the correct offsets for Eastern Time, which is likely to be the
most relevant timezone for the community.

The conversion to local timezone is done in JavaScript so it doesn't
really matter what the offset is so long as the overall time is correct
but thinking in east coast time is probably the mostly likely mindset
for contributors (and causes less friction than requiring all times be
specified in UTC).